+++ crawler/src/main/java/org/kquiet/browserjob/crawler/bigfun/BigFunController.java
@@ -13,8 +13,8 @@ import java.util.stream.Collectors;
 import org.kquiet.browser.ActionComposerBuilder;
 import org.kquiet.browser.BasicActionComposer;
 import org.kquiet.browserjob.crawler.CrawlerBeanConfiguration;
-import org.kquiet.browserjob.crawler.CrawlerService;
 import org.kquiet.browserjob.crawler.bigfun.entity.RealEstate;
+import org.kquiet.browserjob.crawler.service.CrawlerService;
 import org.kquiet.browserscheduler.BrowserSchedulerConfig.JobConfig;
 import org.kquiet.browserscheduler.JobBase;
 import org.openqa.selenium.By;
