+++ crawler/src/main/java/org/kquiet/browserjob/crawler/house/dao/RentHouseRepository.java
@@ -1,6 +1,6 @@
-package org.kquiet.browserjob.crawler.house591.dao;
+package org.kquiet.browserjob.crawler.house.dao;
 
-import org.kquiet.browserjob.crawler.house591.entity.RentHouse;
+import org.kquiet.browserjob.crawler.house.entity.RentHouse;
 import org.springframework.data.jpa.repository.JpaRepository;
 import org.springframework.stereotype.Repository;
