+++ crawler/src/main/resources/application.yml
@@ -65,9 +65,18 @@ browser-scheduler:
     scheduleAfterExec: true
     parameters:
       takeScreenshot: "true"
-  - name: Sale591A
+  - name: SaleYungchingA
     enable: true
-    impl: org.kquiet.browserjob.crawler.house591.Sale591Controller
+    impl: org.kquiet.browserjob.crawler.house.controller.SaleYungchingController
+    dailyStart: "00:00:00"
+    dailyEnd: "00:00:00"
+    interval: 600
+    scheduleAfterExec: true
+    parameters:
+      takeScreenshot: "true"
+  - name: Sale591A
+    enable: false
+    impl: org.kquiet.browserjob.crawler.house.controller.Sale591Controller
     dailyStart: "00:00:00"
     dailyEnd: "00:00:00"
     interval: 600
@@ -76,7 +85,7 @@ browser-scheduler:
       takeScreenshot: "true"
   - name: Rent591A
     enable: false
-    impl: org.kquiet.browserjob.crawler.house591.Rent591Controller
+    impl: org.kquiet.browserjob.crawler.house.controller.Rent591Controller
     dailyStart: "00:00:00"
     dailyEnd: "00:00:00"
     interval: 1800
@@ -85,7 +94,7 @@ browser-scheduler:
       takeScreenshot: "true"
   - name: Rent591B
     enable: false
-    impl: org.kquiet.browserjob.crawler.house591.Rent591Controller
+    impl: org.kquiet.browserjob.crawler.house.controller.Rent591Controller
     dailyStart: "00:00:00"
     dailyEnd: "00:00:00"
     interval: 1800
